+++ manual_tests/demo/legacy_mysql.sql
@@ -1,0 +1,1 @@
+CREATE TABLE articles (id INT, content TEXT) ENGINE=MyISAM;