+++ manual_tests/gen_snowflake.sql
@@ -1,0 +1,2 @@
+-- Migration Script for snowflake
+ALTER TABLE "logs" MODIFY COLUMN "msg" VARCHAR;